punctuator-BiDirectionalLSTM/requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.

tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.7
Werkzeug==1.0.1
wrapt==1.12.1
